requirements.txt
absl-py==0.2.2
algorithmia==1.1.3
astor==0.7.1
certifi==2018.4.16
chardet==3.0.4
enum34==1.1.6
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
idna==2.7
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
Markdown==2.6.11
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.2.0
protobuf==3.6.0
PyYAML==3.13
requests==2.19.1
scipy==1.1.0
six==1.11.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
ujson==1.35
urllib3==1.23
Werkzeug==0.14.1
